Bug 50438 - code calculates average with integer math, expecting double value
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@1044200 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: fb3024aae9f430ef3455735b9a1df7fa3356df13

+++ src/core/org/apache/jmeter/visualizers/SamplingStatCalculator.java
@@ -131,7 +131,7 @@ public class SamplingStatCalculator {
         if (count == 0) {
             return 0;
         }
-        return calculator.getTotalBytes() / count;
+        return calculator.getTotalBytes() / (double) count;
     }
 
     public String getLabel() {
